requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
yarl==1.9.2
